add authorization for recipe/user delete/update

+++ routes/api.php
@@ -10,7 +10,8 @@ use App\Http\Controllers\Api\{
 
 
 use App\Http\Middleware\Auth;
-
+use App\Http\Middleware\AuthorizationUser;
+use App\Http\Middleware\AuthorizationRecipeOwner;
 
 /*
 |--------------------------------------------------------------------------
@@ -28,23 +29,27 @@ use App\Http\Middleware\Auth;
 // }); // auto utworzone
 
 //Routes for Recipe model
-//Route::middleware([Auth::class])->group(function () {
-Route::prefix('/recipes')->group(function () {
-    Route::get('/', [RecipeApiController::class, 'index']);
-    Route::post('/', [RecipeApiController::class, 'store']);
-    Route::get('/{model}', [RecipeApiController::class, 'show']);
-    Route::put('/{model}', [RecipeApiController::class, 'update']);
-    Route::delete('/{model}', [RecipeApiController::class, 'destroy']);
+Route::middleware([Auth::class])->group(function () {
+    Route::prefix('/recipes')->group(function () {
+        Route::get('/', [RecipeApiController::class, 'index']);
+        Route::post('/', [RecipeApiController::class, 'store']);
+        Route::get('/{model}', [RecipeApiController::class, 'show']);
+        Route::put('/{model}', [RecipeApiController::class, 'update'])->middleware(AuthorizationRecipeOwner::class);
+        Route::delete('/{model}', [RecipeApiController::class, 'destroy'])->middleware(AuthorizationRecipeOwner::class);
+    });
 });
-//});
 
 // Routes for User model
-Route::prefix('/users')->group(function () {
-    Route::get('/', [UserApiController::class, 'index']);
-    Route::post('/', [UserApiController::class, 'store']);
-    Route::get('/{model}', [UserApiController::class, 'show']);
-    Route::put('/{model}', [UserApiController::class, 'update']);
-    Route::delete('/{model}', [UserApiController::class, 'destroy']);
+Route::middleware([Auth::class])->group(function () {
+    Route::prefix('/users')->group(function () {
+        Route::get('/', [UserApiController::class, 'index']);
+        Route::post('/', [UserApiController::class, 'store']);
+        Route::get('/{model}', [UserApiController::class, 'show']);
+        Route::put('/{model}', [UserApiController::class, 'update'])->middleware(AuthorizationUser::class);
+        ;
+        Route::delete('/{model}', [UserApiController::class, 'destroy'])->middleware(AuthorizationUser::class);
+        ;
+    });
 });
 
 Route::prefix('/user-access-tokens')->group(function () {
